// types.ts
export type TreeState = 'CHAOS' | 'FORMED';

export interface TreeColors {
  bottom: string;
  top: string;
}

export interface HandGesture {
  isOpen: boolean;
  position: { x: number; y: number }; // Normalized -1 to 1
  isDetected: boolean;
}

export type LuckyDrawStatus = 'IDLE' | 'PICKING' | 'WINNER' | 'REVEALED';